chore: reestructure project
