+++ .gitignore
@@ -1,7 +1,5 @@
+__pycache__/
 
-/.venv
-/__pycache__
-/API/__pycache__
-/databaseHandling/__pycache__
-/flask_session
-images/
+.venv
+flask_session
+logos
