--- databaseHandling/README.md
@@ -1,5 +1,0 @@
-# What's in here?
-> This folder contains all the files to access/manage the database. There is a test file for you to see the basic logic of how we implement the calls, it's called `database_test_maisn.py`.
-
-# What to do here?
-The databse is almost complete, I would like to have a few extra functions, regarding `member.py`, that would allow us to manage/return logos and edit password, as well as adding encryption to this passwords.
