+++ src/api/member_projects/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('member_projects', __name__)
+
+from api.member_projects import routes
