+++ src/api/members/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('members', __name__)
+
+from api.members import routes
