+++ src/api/projects/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('projects', __name__)
+
+from api.projects import routes
