+++ src/services/logo.py
